+++ src/app/categories/categories.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import {CategoriesStorageService} from "../services/categories-storage.service";
 
 @Component({
   selector: 'categories',
@@ -6,10 +7,19 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./categories.component.css']
 })
 export class CategoriesComponent implements OnInit {
+  categories : string[] = [];
+  newCategory: string = '';
 
-  constructor() { }
+  constructor(private categoriesService: CategoriesStorageService) { }
 
   ngOnInit(): void {
+    this.categoriesService.$categories.subscribe(categories => this.categories = categories);
+    this.categoriesService.save(['bla', 'bla12', 'alsdkfja;'])
   }
 
+  addCategory(value: string) {
+    this.categoriesService.add(value);
+
+    this.newCategory = '';
+  }
 }