+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - styles for Savings
 - styles for Spends
 - modals for buttons
+- easy category functionality
 
 ### Fixed
 - back favicon